pipeline/requirements.txt
anthropic
supabase
feedparser
requests
python-dotenv